+++ src/verisocks.c
@@ -34,11 +34,11 @@ SOFTWARE.
 #include <string.h>
 
 #include "verisocks.h"
+#include "version.h"
 #include "vs_logging.h"
 #include "vs_utils.h"
 #include "vs_server.h"
 #include "vs_msg.h"
-#include "vs_itx.h"
 #include "vs_vpi.h"
 
 #define READ_BUFFER_SIZE 4096
@@ -210,12 +210,11 @@ PLI_INT32 verisocks_init_calltf(PLI_BYTE8 *user_data)
     p_vpi_data->fd_client_socket = -1;
     p_vpi_data->p_cmd = NULL;
     p_vpi_data->h_cb = NULL;
+    p_vpi_data->h_cb_timeout = NULL;
     p_vpi_data->value = default_value;
     p_vpi_data->uuid.valid = 0u;
     p_vpi_data->time_def = vs_utils_get_sim_time_def();
     p_vpi_data->sim_time = 0ull;
-    vs_vpi_itx_t itx_init_table[VS_VPI_MAX_ITX] = {VS_VPI_ITX_NULL};
-    memcpy(p_vpi_data->itx_table, itx_init_table, sizeof(itx_init_table));
     memcpy(&p_vpi_data->uuid.value, null_uuid_value, VS_UUID_LEN);
     vpi_put_userdata(h_systf, (void*) p_vpi_data);
 
@@ -243,6 +242,7 @@ PLI_INT32 verisocks_init_calltf(PLI_BYTE8 *user_data)
     vpi_printf("*                                        *\n");
     vpi_printf("******************************************\n");
 
+    vs_vpi_log_info("Verisocks version: " VERISOCKS_VERSION);
     vs_vpi_log_info("Server address: %d.%d.%d.%d",
         (s_addr & 0xff000000) >> 24u,
         (s_addr & 0x00ff0000) >> 16u,
@@ -310,6 +310,12 @@ PLI_INT32 verisocks_cb(p_cb_data cb_data)
         p_vpi_data->h_cb = NULL;
     }
 
+    /* If there is a timeout callback, remove it */
+    if (NULL != p_vpi_data->h_cb_timeout) {
+        vpi_remove_cb(p_vpi_data->h_cb_timeout);
+        p_vpi_data->h_cb_timeout = NULL;
+    }
+
     /* Check state */
     if (p_vpi_data->state != VS_VPI_STATE_SIM_RUNNING) {
         vs_vpi_log_error("Inconsistent state");
@@ -349,7 +355,73 @@ for command ...");
 }
 
 /**
- * @brief Callback function - Used for for_until_change
+ * @brief Callback function - Used for until_change timeout
+ *
+ * @param cb_data Pointer to s_cb_data struct
+ * @return Returns 0 if successful, -1 in case of error
+ */
+PLI_INT32 verisocks_cb_timeout(p_cb_data cb_data)
+{
+    /* Retrieve stored user data */
+    vs_vpi_data_t *p_vpi_data = NULL;
+    p_vpi_data = (vs_vpi_data_t*) cb_data->user_data;
+    if (NULL == p_vpi_data) {
+        vs_vpi_log_error("Could not get stored data - Aborting callback");
+        goto error;
+    }
+
+    /* Free callback handle */
+    if (NULL != p_vpi_data->h_cb_timeout) {
+        vpi_free_object(p_vpi_data->h_cb_timeout);
+        p_vpi_data->h_cb_timeout = NULL;
+    }
+
+    /* If there is a timeout callback, remove it */
+    if (NULL != p_vpi_data->h_cb) {
+        vpi_remove_cb(p_vpi_data->h_cb);
+        p_vpi_data->h_cb = NULL;
+    }
+
+    /* Check state */
+    if (p_vpi_data->state != VS_VPI_STATE_SIM_RUNNING) {
+        vs_vpi_log_error("Inconsistent state");
+        goto error;
+    }
+
+    /* Update sim time state variable */
+    p_vpi_data->sim_time = vs_utils_get_sim_time(p_vpi_data->time_def);
+
+    /* Signalling that the timeout callback function has been reached */
+    vs_vpi_log_info("Reached timeout callback - \
+Verisocks taking over and waiting for command ...");
+    VS_VPI_RETURN(p_vpi_data, "timeout",
+        "Reached timeout callback - Getting back to Verisocks main loop");
+
+    /* Call verisocks main loop */
+    p_vpi_data->state = VS_VPI_STATE_WAITING;
+    if (0 > verisocks_main(p_vpi_data)) {
+        goto error;
+    }
+    vs_vpi_log_info("Releasing control to simulator");
+    return 0;
+
+    /* Error management */
+    error:
+    if (NULL != p_vpi_data) {
+        p_vpi_data->state = VS_VPI_STATE_ERROR;
+        if (0 <= p_vpi_data->fd_server_socket) {
+            close(p_vpi_data->fd_server_socket);
+            p_vpi_data->fd_server_socket = -1;
+        }
+        //free(p_vpi_data);  //Will be freed in exit callback handler
+    }
+    vs_vpi_log_info("Aborting simulation");
+    vpi_control(vpiFinish, 1);
+    return -1;
+}
+
+/**
+ * @brief Callback function - Used for until_change
  *
  * @param cb_data Pointer to s_cb_data struct
  * @return Returns 0 if successful, -1 in case of error
@@ -374,7 +446,8 @@ PLI_INT32 verisocks_cb_value_change(p_cb_data cb_data)
     p_vpi_data->sim_time = vs_utils_get_sim_time(p_vpi_data->time_def);
 
     /* If the value is not the same, get back to sim until next time */
-    if (vpi_get(vpiType, cb_data->obj) != vpiNamedEvent) {
+    // if (vpi_get(vpiType, cb_data->obj) != vpiNamedEvent) {
+    if (vpiSuppressVal != p_vpi_data->value.format) {
         if (vs_utils_compare_values(
             p_vpi_data->value, *(cb_data->value)) != 0) {
             return 0;
@@ -612,93 +685,4 @@ static PLI_INT32 verisocks_main_waiting(vs_vpi_data_t *p_vpi_data)
 valid JSON content. Discarding it.");
     VS_VPI_RETURN(p_vpi_data, "error", "Invalid message content - Discarding");
     return 0;
-}
-
-PLI_INT32 verisocks_cb_itx(p_cb_data cb_data)
-{
-    vs_vpi_itx_t *p_itx = NULL;
-    vs_vpi_data_t *p_vpi_data = NULL;
-
-    /* Retrieve stored user ITX data */
-    p_itx = (vs_vpi_itx_t*) cb_data->user_data;
-    if (NULL == p_itx) {
-        vs_vpi_log_warning( 
-            "Could not get stored PLI ITX data - Aborting ITX callback");
-        /* Return to simulation without doing anything */
-        return 0;
-    }
-
-    /* Retrieve pointer to PLI application data */
-    p_vpi_data = (vs_vpi_data_t*) p_itx->user_data;
-    if (NULL == p_itx) {
-        vs_vpi_log_warning( 
-            "Could not get stored PLI ITX data - Aborting ITX callback");
-        /* Return to simulation without doing anything */
-        return 0;
-    }
-
-    /* Check state consistency */
-    if (p_vpi_data->state != VS_VPI_STATE_SIM_RUNNING) {
-        vs_vpi_log_error("Inconsistent state in callback handler");
-        return -1;
-    }
-
-    /* Update sim time state variable */
-    p_vpi_data->sim_time = vs_utils_get_sim_time(p_vpi_data->time_def);
-
-    /* Logging that the callback function has been reached */
-    vs_vpi_log_info("ITX %s callback reached", p_itx->name);
-
-    /* Behavior dependent on interrupt type */
-    if (p_itx->type == VS_VPI_ITX_IN_TIME) {
-
-        /* Free callback handle */
-        if (NULL != p_itx->h_cb) {
-            vpi_free_object(p_itx->h_cb);
-            p_itx->h_cb = NULL;
-        }
-
-        /* If the ITX is recurrent ... */
-        if (itx_is_recurrent(p_itx)) {
-            /* Re-register the exact same callback, thus we keep it active and
-            just re-create and update the handle */
-            vpiHandle h_cb;
-            h_cb = vpi_register_cb(cb_data);
-            p_itx->h_cb = h_cb;
-        }
-
-        /* Blocking interrupt */
-        if (itx_is_blocking(p_itx)) {
-            /* Recurrent */
-            p_vpi_data->state = VS_VPI_STATE_SIM_BLOCKED;
-            vs_vpi_itx_return(p_itx);
-            vs_vpi_log_info("Blocking ITX - Verisocks taking over and waiting \
-for command ...");
-            if (0 > verisocks_main(p_vpi_data)) {
-                return -1;
-            }
-            vs_vpi_log_info("Releasing control back to simulator");
-            return 0;
-        }
-        /* Non-blocking interrupt */
-        else {
-        }
-    }
-    else if (p_itx->type == VS_VPI_ITX_ON_CHANGE) {
-        /* If recurrent */
-        /* Non-blocking interrupt */
-        /* Blocking interrupt - Considered as non-recurrent */
-    }
-    else if (p_itx->type == VS_VPI_ITX_AT_TIME) {
-        /* Non-blocking interrupt */
-        /* Blocking interrupt */
-    }
-
-    /* Free callback handle */
-    if (NULL != p_itx->h_cb) {
-        vpi_remove_cb(p_itx->h_cb);
-        p_itx->h_cb = NULL;
-    }
-    // error:
-    return -1;
 }